+++ src/app/components/dynamic/translatable-string/translatable-string.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 import { SimplifiedTranslationService } from 'src/app/services/simplifiedTranslation/simplified-translation.service';
 
 @Component({
@@ -9,14 +11,14 @@ import { SimplifiedTranslationService } from 'src/app/services/simplifiedTransla
 export class TranslatableStringComponent implements OnInit {
 
   @Input() text: string;
+  text$: Observable<string>;
 
   constructor(private translator: SimplifiedTranslationService) {}
 
   ngOnInit(): void {
-    this.text = this.translator.syncTranslate(this.text);
-    this.translator.getCurrentLang().subscribe(lang => {
-      this.text = this.translator.syncTranslate(this.text);
-    });
+    this.text$ = this.translator.getCurrentLang().pipe(map((l) => {
+      return this.translator.syncTranslate(this.text);
+    }));
   }
 
 }